move youtube client to server code
