+++ src/lib/services/server/youtube/index.ts
@@ -5,4 +5,3 @@ export {
   type Thumbnails,
   type Thumbnail,
 } from "./getYoutubeVideoData";
-export { default as getYoutubeIDFromURL } from "./getYoutubeIDFromURL";
